refactor: Move Content context provider to pages.
