+++ src/admin/pages/collections/List/Singleton.tsx
@@ -1,6 +1,6 @@
 import RenderFields from '@admin/components/forms/RenderFields';
 import ComposeWrapper from '@admin/components/utilities/ComposeWrapper';
-import { ContentContextProvider, useContent } from '@admin/stores/Content';
+import { ContentContextProvider, useContent } from '@admin/pages/collections/Context';
 import { Button, Drawer, Stack, useTheme } from '@mui/material';
 import Grid from '@mui/material/Unstable_Grid2';
 import { useSnackbar } from 'notistack';
